Confetti added to doot animation
