+++ package.json
@@ -33,6 +33,7 @@
     "@testing-library/user-event": "^14.5.2",
     "@types/bcryptjs": "^2.4.6",
     "@types/better-sqlite3": "^7.6.10",
+    "@types/canvas-confetti": "^1.6.4",
     "@types/cookie": "^0.6.0",
     "@types/crypto-js": "^4.2.2",
     "@types/fs-extra": "^11.0.4",
@@ -76,6 +77,7 @@
     "@prisma/client": "^5.20.0",
     "bcryptjs": "^2.4.3",
     "better-sqlite3": "^9.3.0",
+    "canvas-confetti": "^1.9.3",
     "classnames": "^2.5.1",
     "close-with-grace": "^1.3.0",
     "cross-env": "^7.0.3",
