+++ src/lib/components/svgs/Logo/Logo.svelte
@@ -1,138 +1,21 @@
 <script lang="ts">
-	import gsap from 'gsap'
-	import { CustomEase } from 'gsap/dist/CustomEase'
-	import { browser } from '$app/environment'
+	import {
+		animateLogo,
+		createTimeline,
+		getDurations,
+		getMarkers,
+		initializeGSAP,
+	} from './helpers.logo'
 
-	let isJsEnabled = $state(false)
+	let logo: SVGElement | null = $state(null)
 
-	// Base animation durations
-	const CONFIG = {
-		animation: {
-			fullBounceTime: 1.5,
-			honkTime: 0.6,
-			fullXOffset: 200,
-		},
-		ratios: {
-			peak1: 0.4, // 40% of fullBounceTime
-			groundBounce: 0.6, // 60% of fullBounceTime
-			peak2: 0.8, // 80% of fullBounceTime
-			honkOffset: 0.2,
-		},
-		customEase: {
-			name: 'bouncy',
-			path: 'M0,0 C0.045,0.401 0.178,1.412 0.373,1.412 0.514,1.412 0.568,1.19 0.599,1.014 0.634,1.191 0.711,1.217 0.804,1.217 0.95,1.217 1,1 1,1',
-		},
-	}
-
-	// Computed timing markers based on fullBounceTime
-	const getMarkers = (fullBounceTime: number, honkTime: number) => ({
-		start: 0,
-		peak1: CONFIG.ratios.peak1 * fullBounceTime,
-		groundBounce: CONFIG.ratios.groundBounce * fullBounceTime,
-		peak2: CONFIG.ratios.peak2 * fullBounceTime,
-	})
-
-	// Computed durations for animations
-	const getDurations = (markers: ReturnType<typeof getMarkers>) => ({
-		initUp: markers.peak1,
-		initDown: markers.groundBounce - markers.peak1,
-		secondUp: markers.peak2 - markers.groundBounce,
-		secondDown: CONFIG.animation.fullBounceTime - markers.peak2,
-	})
-
-	function initializeGSAP() {
-		gsap.registerPlugin(CustomEase)
-		CustomEase.create(CONFIG.customEase.name, CONFIG.customEase.path)
-
-		gsap.set('#logo', {
-			y: 140,
-			rotate: 35,
-			opacity: 0,
-		})
-	}
-
-	// Main animation setup
 	$effect(() => {
-		if (browser) {
-			isJsEnabled = true
-
-			initializeGSAP()
-			const markers = getMarkers(
-				CONFIG.animation.fullBounceTime,
-				CONFIG.animation.honkTime,
-			)
-			const durations = getDurations(markers)
-			const tl = gsap.timeline()
-
-			for (const [key, value] of Object.entries(markers)) {
-				tl.addLabel(key, value)
-			}
-
-			tl.to('#logo', {
-				y: 0,
-				duration: CONFIG.animation.fullBounceTime,
-				ease: CONFIG.customEase.name,
-			})
-				.to(
-					'#logo',
-					{
-						duration: durations.initUp,
-						rotate: 0,
-						opacity: 1,
-						x: CONFIG.animation.fullXOffset * CONFIG.ratios.peak1,
-						ease: 'none',
-					},
-					'start',
-				)
-				.to(
-					'#logo',
-					{
-						duration: durations.initDown,
-						rotate: -20,
-						x: CONFIG.animation.fullXOffset * CONFIG.ratios.groundBounce,
-						ease: 'none',
-					},
-					'peak1',
-				)
-				.to(
-					'#logo',
-					{
-						duration: durations.secondUp,
-						rotate: 15,
-						x: CONFIG.animation.fullXOffset * CONFIG.ratios.peak2,
-						ease: 'none',
-					},
-					'groundBounce',
-				)
-				.to(
-					'#logo',
-					{
-						duration: durations.secondDown,
-						rotate: -10,
-						x: CONFIG.animation.fullXOffset,
-						ease: 'none',
-					},
-					'peak2',
-				)
-				.to('#logo', {
-					duration: CONFIG.animation.honkTime,
-					x: 100,
-					y: 50,
-					rotate: 15,
-					ease: 'expo.out',
-				})
-				.to(
-					'#logo',
-					{
-						duration: CONFIG.animation.honkTime,
-						x: 0,
-						y: 0,
-						rotate: -15,
-						ease: 'expo.out',
-					},
-					`-=${CONFIG.ratios.honkOffset}`,
-				)
-		}
+		initializeGSAP()
+		const markers = getMarkers()
+		const durations = getDurations(markers)
+		const tl = createTimeline(markers)
+		
+		animateLogo(tl, durations)
 	})
 </script>
 
@@ -141,6 +24,9 @@
 	id="logo"
 	viewBox="0 0 205.826 81.578"
 	xmlns="http://www.w3.org/2000/svg"
+	aria-label="Animated logo"
+	role="img"
+	bind:this={logo}
 >
 	<path
 		d="M0 0c.926.035 1.852.066 2.773.094C1.336-2.043.367-4.414-.269-6.898-.234-4.602-.133-2.305 0 0Zm758.063-127.793c10.62 49.855 5.55 101.805-7.383 150.73-7.211 32.493-14.504 64.973-23.309 97.082-8.219 25.614-18.453 79.028-53.101 76.227-41.391-3.988-81.793-26.203-103.465-62.289-23.95-38.2-34.473-43.555-77.02-56.23 5.813 34.09-.25 61.136-14.363 81.375v.004C462.5 183.37 434 197.844 400.516 202.938h-.012c-15.016 2.285-31.035 2.68-47.465 1.222a480.253 480.253 0 0 0-4.621-.414c-39.887-3.41-79.961-2.265-119.926-.738-48.652 1.285-116.293.598-154.762-32.707-19.601-18.817-37.851-44.043-45.82-70.184-3.894-16.316-6.894-33.023-8.355-49.816a249.905 249.905 0 0 1-.98-22.344v-.187c.003-.93.01-1.856.023-2.786-1.032.004-2.063.004-3.094-.004a300.291 300.291 0 0 1-6.223-.093c-7.125-.157-14.254-.473-21.379-.606-3.03-.449-14.543 1.988-10.671-4.12 5.261-2.583 12.007-.763 17.785-1.524 3.199-.09 6.394-.176 9.593-.278a13.24 13.24 0 0 1-1.14-1.574c-11.371-.023-19.375-2.984-18.48-15.953.468-7.906 2.398-20.945 7.992-28.734-7.485-.707-14.973-1.414-22.481-1.742-8.195-.137-8.223-12.465 0-12.579 13.285-.414 26.543.035 39.809.41 0 .864.011 1.723.043 2.583-.102 18.386.847 36.796-.11 55.167-.015.235-.035.465-.05.7 2.66-.781 5.277-1.43 7.41-1.922h.003a243.95 243.95 0 0 1 1.668-.375c.106-.024.207-.047.305-.067a17.8 17.8 0 0 0 .371-.082c.086-.02.172-.039.258-.054a141.68 141.68 0 0 1 1.48-.313l.172-.035a195.202 195.202 0 0 1 4.36-.777l.781-.13c.602-.097 1.203-.195 1.809-.284.113-.02.226-.04.34-.055a78.544 78.544 0 0 1 1.64-.242c.012 0 .023-.004.035-.004.38-.059.754-.113 1.133-.164.55-.078 1.102-.156 1.652-.227.192-.023.38-.05.57-.074.536-.07 1.067-.137 1.602-.2.356-.046.711-.089 1.07-.128l.086-.012a105.584 105.584 0 0 1 1.895-.215c.484-.054.965-.105 1.45-.152.253-.031.511-.055.765-.082 3.867-.387 7.746-.68 11.625-.898 20.027-1.07 39.926 2.003 59.828 3.687 10.547.898 21.117.695 31.68 1.09 8.691.574 17.332 1.703 26.015 2.344 30.504 1.504 61.067.601 91.594.757 25.977-.523 52.016-1.257 77.82 2.387 15.04 1.496 30.543 2.293 44.254 9.3 4.86 2.185 9.91 4.786 13.368 8.954 6.132 7.34 9.492 16.918 9.234 26.477-.227 1.394-.45 2.785-.672 4.172-.762 4.023-2.648 9.183-3.496 14.113-7.766 17.527-23.234 24-41.824 28.465-15.325 4.55-31.391 4.796-47.211 3.515-5.192-.34-10.387-.707-15.578-1.078 15.699.524 31.375.61 46.886-2.062 22.29-4.907 57.485-12.93 44.465-43.325-53.715-3.23-107.23 1.922-160.476 8.176-.278.043-.555.086-.844.125h-.004c-17.996 2.516-55.895-2.14-68.317 7.418 0 .004-.07-.004-.07-.004-1.91 3.375-1.433 6.418.809 9.145 7.347 8.93 33.672 14.48 57.332 17.375-1.004.031-2.012.066-3.016.105-20.633.332-42.074 2.84-61.746-4.933-10.152-3.227-16.535-13.871-14.445-24.352 2.937-13.719 18.668-17.402 30.738-18.566 10.793-1.266 21.66-1.536 32.5-2.157 15.54-1.074 31.121-1.82 46.7-1.586 27.948.313 55.656 4.54 83.585 5.297 20.656 1.653 41.23-3.636 61.871-.996.078-8.133-3.726-17.183-11.515-20.601-12.301-6.38-25.93-8.317-39.578-9.39-12.512-1.298-25.04-2.829-37.645-2.759-25.344.18-50.687 1.047-76.04.32-27.14-.367-54.272-1.14-81.312-3.617-12.628-.613-25.285.371-37.93.239-19.519-.282-38.948 2.64-58.488 2.18a257.09 257.09 0 0 1-9.449-.434c0 .058-.004.12-.004.18-.007.253-.011.503-.02.757-.007.14-.01.285-.015.43-.008.262-.02.527-.031.789-.035.79-.078 1.582-.14 2.367-.024.262-.044.524-.071.785-.012.13-.023.262-.04.391-.026.277-.058.555-.093.828-.031.242-.062.484-.097.727-.008.062-.02.129-.028.191.004.047.008.09.008.137.059.894.113 1.789.164 2.683.055.891.102 1.778.152 2.664 0 .036.004.07.004.106.047.867.09 1.734.13 2.601.089 1.793.167 3.586.241 5.38.04.894.079 1.788.114 2.687.054 1.32.105 2.644.16 3.965.031.843.066 1.687.101 2.531.036.844.067 1.684.106 2.527.058 1.473.125 2.95.195 4.422.024.563.05 1.125.082 1.688.024.492.047.984.074 1.476a565.563 565.563 0 0 0 .277 4.895c.055.894.114 1.789.177 2.683.84 17.836 2 36.504 12.484 51.672 4.766 6.918 10.68 12.89 16.46 18.942 5.458 5.964 11.2 11.683 17.981 16.152 9.489 6.168 20.637 9.312 31.508 12.09 17.278 2.601 34.797 3.031 52.242 3.015 20.973.118 41.946-.421 62.848-2.222 14.96-.848 29.953-.535 44.934-.711 29.16-.844 58.285-.168 87.425.926 18.782.03 37.641-1.168 56.172-4.23 8.235-1.99 16.7-4.556 23.399-9.919 3.652-3.156 5.746-7.586 7.941-11.789 5.805-11.73 11.656-23.484 16.43-35.676 2.965-9.277 4.199-18.976 4.984-28.656.738-11.75-.762-23.441-2.535-35.027-1.586-10.696-1.95-21.871-6.727-31.766-2.035-4.039-4.68-7.707-7.644-11.105-6.578-7.352-13.453-14.606-21.598-20.243-10.562-7.16-22.55-11.761-34.418-16.261-11.261-3.617-23.097-4.977-34.777-6.543-18.816-2.438-36.75-3.309-55.469-3.446-41.43-1.015-82.87-1.269-124.3-2.21-19.665.097-39.305-1.329-58.95-1.485-9.89.086-19.773.078-29.656-.312-1.04-.051-2.082-.106-3.121-.164-.555-.028-1.106-.059-1.66-.09-.801-.043-1.598-.09-2.399-.137-.543-.035-1.082-.066-1.62-.098a963.736 963.736 0 0 1-7.395-.488c-1.278-.086-2.555-.18-3.832-.273-.121-.008-.243-.016-.364-.028a631.29 631.29 0 0 1-3.898-.293c-1.352-.101-2.7-.207-4.047-.316-.965-.078-1.93-.152-2.89-.234-1.231-.098-2.462-.2-3.692-.305-1.711-.14-3.418-.29-5.125-.434l-2.86-.246c-3.812-.332-7.62-.668-11.433-1l-4.805-.422c-1.867-.164-3.738-.324-5.605-.484a90.515 90.515 0 0 1 6.199-8.453c15.77-19.137 38.64-31.664 61.703-40.383 22.684-9.379 46.02-17.55 70.246-21.863 39.035-6.703 78.63-.258 117.82 1.308 59.54 4.477 119.274 9.63 178.997 8.934 57.363-5.691 124.39-6.277 166.73-51.16 18.961-19.7 31.7-58.836 59.906-65.055 37.528-3.215 53.465 66.68 59.332 94.879z"
@@ -187,17 +73,11 @@
 </svg>
 
 <style>
-	/* Initial state (pre-animation) */
 	#logo {
 		transform: rotate(-15deg);
+		will-change: transform, opacity;
 	}
 
-	/* JS enabled - let GSAP take over */
-	#logo.js-enabled {
-		transition: none; /* Prevents CSS transitions from interfering with GSAP */
-	}
-
-	/* No-JS fallback - show final state */
 	@media (prefers-reduced-motion: reduce), (script: none) {
 		#logo {
 			transform: translate(0, 0) rotate(-15deg);
